tests: Move pod servers to libnet
The servers were part of the top-level tests package. However,
functionally they rather belong to libnet.
