+++ tests/libnet/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "network_attachment_definition.go",
         "ping.go",
         "plumbing.go",
+        "servers.go",
         "skips.go",
         "validation.go",
         "vmibuilder.go",
@@ -24,6 +25,8 @@ go_library(
         "//tests/framework/kubevirt:go_default_library",
         "//tests/libnet/cloudinit:go_default_library",
         "//tests/libnet/cluster:go_default_library",
+        "//tests/libpod:go_default_library",
+        "//tests/testsuite:go_default_library",
         "//vendor/github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
